File management update
First version was missing the Common folder, so that's now been added.  Another change is that the HLSL has been excluded from the build. I have also had to fix the headers as they were searching in the wrong part of the directory from back when the common was outside the main folder. The project now runs again in the repo.