Fixes to Shader, and Vertex and Index Buffers
The previous version had an issue with the vertex and index buffers. This was discovered to be an issue where the points in memory were incorrent. This has been adjusted so now the scene is not only showing the shapes and their colours, but allowing for more than one piece of geormetry if needed. The next issue was with the shader, in that it was still treating the vertex variables as two seperate buffers. This has been resolved, and is now working with the one.