Name Changes
The names of files have been updated to provide a clearer look at the application. Originally the main files were names Box, due to the base version being an application that draws a box. Now that this is updated to App, and the solution to Honours Project, it reflects the contents better.