Model Loading Working
TinyObj is now working (at least for models). As always the issue was rediculously silly. Kicking myself right now. Make sure that the data sizes are logical, as it leads to too many issues. Next step could either be texturing (shouldn't be hard) or improving how the particles are handled.