Multiple Particles
The scene now renders multiple particles. The issue is the current implementation is slow with a lot of particles (around 100). This is likely due to how they are updating, and all having there own mesh geometry. Going to try bundling them up in a single one for the next step. See if that cuts time down at all.